pyproject.toml
[tool.poetry]
name = "flow-gtd"
version = "0.1.0"
description = "Local-First, AI-Augmented GTD CLI for Senior Engineering Managers"
authors = ["Flow Team"]
readme = "README.md"
packages = [{include = "flow"}]

[tool.poetry.scripts]
flow = "flow.main:main"

[tool.poetry.dependencies]
python = "^3.11"

# CLI Layer
typer = {extras = ["all"], version = "^0.15.0"}

# TUI Layer
textual = "^3.0.0"

# LLM Integration
google-genai = "^1.0.0"

# RAG / Embeddings
sentence-transformers = "^3.0.0"
chromadb = "^0.6.0"

# Configuration
pydantic-settings = "^2.7.0"

# Apple Ecosystem Sync (macOS only)
pyobjc-framework-EventKit = {version = "^11.0", markers = "sys_platform == 'darwin'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
textual-dev = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
